requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
dataclasses-json==0.6.1
Django==4.2
dnspython==2.4.2
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.317
langsmith==0.0.46
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
pandas==2.1.1
pinecone-client==2.2.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
win32-setctime==1.1.0
yarl==1.9.2
